move components into their file, set a global setting
