+++ frontend/src/src/components/LayerSelection.svelte
@@ -1,8 +1,9 @@
 <script>
 	import { onMount } from 'svelte';
-	import './L.TileLayer.NutsLayer.js'
-	import './L.DrawingLayer.js'
+	import '../leaflet_components/L.TileLayer.NutsLayer.js'
+	import '../leaflet_components/L.DrawingLayer.js'
 	import queryString from "query-string";
+	import { BASE_URL } from '../settings.js'
 
 	export const SELECTIONS = new Set(['lau.zip', 'nuts0.zip', 'nuts1.zip', 'nuts2.zip' ,'nuts3.zip'])
 	let selection_layers = [];
@@ -63,7 +64,7 @@
 	}
 	async function fetchLayers() {
 		//document.domain = "geoserver.hotmaps.eu";
-		let response = await fetch('/api/geofile');
+		let response = await fetch(BASE_URL + 'api/geofile');
 		if (!response.ok) {
 			console.log(response);
 			return [];
@@ -73,7 +74,7 @@
 	}
 	function to_leaflet_layer(layer_name) {
 		const layer =  L.tileLayer.nutsLayer(
-		'/api/wms?',
+		BASE_URL + 'api/wms?',
 		{
 		transparent: 'true',
 		layers: layer_name,
