+++ frontend/src/src/leaflet_components/L.DrawingLayer.js
