+++ frontend/src/src/settings.js
@@ -1,0 +1,1 @@
+export const BASE_URL = "http://127.0.0.1:7000/"
